% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{euroleague_basketball}
\alias{euroleague_basketball}
\title{Euroleague Basketball Teams Data}
\format{
A data.table with 20 rows and 12 variables:
\describe{
\item{Team}{Character. Team name.}
\item{\verb{Home city}}{Character. Home city.}
\item{Arena}{Character. Arena(s) used.}
\item{Capacity}{Character. Stated capacity (may contain multiple values).}
\item{\verb{Last season}}{Character. Last EuroLeague season finish.}
\item{Country}{Character. Country.}
\item{FinalFour_Appearances}{Character. Number of Final Four appearances.}
\item{Titles_Won}{Character. Number of EuroLeague titles.}
\item{Years_of_FinalFour_Appearances}{Character. Years of Final Four appearances (comma-separated).}
\item{Years_of_Titles_Won}{Character. Years of titles (comma-separated).}
}
}
\source{
Wikipedia: EuroLeague page; merged with your curated finals file.
See \url{https://en.wikipedia.org/wiki/EuroLeague} and this package's
source \code{data-raw/} scripts.
}
\usage{
euroleague_basketball
}
\description{
Cleaned information on Euroleague teams: home cities, arenas, capacities,
and historical performance (Final Four appearances and titles).
}
\keyword{datasets}
