% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{euroleague_basketball}
\alias{euroleague_basketball}
\title{Euroleague Basketball Teams Data}
\format{
\subsection{\code{euroleague_basketball}}{

A data frame with 23 rows and 11 columns:
\describe{
\item{Team}{Name of the basketball team}
\item{Home city}{City where the team is based}
\item{Arena}{Name of the team's arena}
\item{Capacity}{Seating capacity of the arena}
\item{Season}{The team's performance in the current season}
\item{Last season}{The team's performance in the last season}
\item{Country}{Country where the team is located}
\item{FinalFour_Appearances}{Number of times the team has reached the Euroleague Final Four}
\item{Titles_Won}{Number of Euroleague titles won by the team}
\item{Years_of_FinalFour_Appearances}{Years when the team reached the Final Four}
\item{Years_of_Titles_Won}{Years when the team won the Euroleague title}
}
}
}
\source{
Data compiled from Euroleague basketball statistics.
}
\usage{
euroleague_basketball
}
\description{
A dataset containing information about Euroleague basketball teams, their home cities, arenas, and performance in the competition.
}
\keyword{datasets}
